refactor ast_parser, parse ast item seperately
