+++ src/parser/ast_parser.rs
@@ -1,228 +1,9 @@
 // The basic idea is, extend crates ra_ap_syntax, especially ast, to support UML entity methods.
-use ra_ap_syntax::{ast::{self, AstNode, HasName, HasModuleItem}, match_ast, SourceFile, Parse};
+use ra_ap_syntax::{ast::{self, HasModuleItem}, SourceFile, Parse};
 
 use crate::uml_entity::*;
 use super::StringParser;
-
-trait HasUMLEntity {
-    fn get_uml_entities(&self) -> Vec<UMLEntity>; // get uml entities from all types of ast entities
-}
-
-impl HasUMLEntity for ast::Struct {
-    fn get_uml_entities(&self) -> Vec<UMLEntity> {
-        let mut results = vec![];
-        let mut record_fields = vec![];
-        for node in self.syntax().descendants() {
-            match_ast! {
-                match node {
-                    ast::RecordField(rf) => {
-                        // get fields for UMLClass
-                        record_fields.push(rf.to_string());
-
-                        // get Aggregation and Composition Relations
-                        let rf_str = rf.to_string();
-                        if rf_str.contains(r"*mut") || rf_str.contains(r"*const") {
-                            get_paths_str_from_ast_node(rf)
-                                .iter()
-                                .for_each(|p| results.push(
-                                    UMLEntity::UMLRelation(UMLRelation::new(&self.name().unwrap().text().to_string(), &p, UMLRelationKind::UMLAggregation)))
-                                )
-                        } else if !rf_str.contains(r"*mut") && !rf_str.contains(r"*const") {
-                            get_paths_str_from_ast_node(rf)
-                                .iter()
-                                .for_each(|p| results.push(
-                                    UMLEntity::UMLRelation(UMLRelation::new(&self.name().unwrap().text().to_string(), &p, UMLRelationKind::UMLComposition)))
-                                )
-                        }
-                    },
-                    _ => ()
-                }
-            }
-            // println!("{:?}", node);
-            // println!("{}", node);
-        };
-        results.push(UMLEntity::UMLClass(UMLClass::new(&self.name().unwrap().text().to_string(), record_fields, vec![], UMLClassKind::UMLClass)));
-        results
-    }
-}
-
-impl HasUMLEntity for ast::Trait {
-    fn get_uml_entities(&self) -> Vec<UMLEntity> {
-        let mut results = vec![];
-        // add UMLClass
-        results.push(UMLEntity::UMLClass(UMLClass::new(&self.name().unwrap().text().to_string(), vec![], vec![], UMLClassKind::UMLTrait)));
-
-        for node in self.syntax().descendants() {
-            match_ast! {
-                match node {
-                    ast::RecordField(rf) => {
-                        // get Aggregation and Composition Relations
-                        let rf_str = rf.to_string();
-                        if rf_str.contains(r"*mut") || rf_str.contains(r"*const") {
-                            get_paths_str_from_ast_node(rf)
-                                .iter()
-                                .for_each(|p| results.push(
-                                    UMLEntity::UMLRelation(UMLRelation::new(&self.name().unwrap().text().to_string(), &p, UMLRelationKind::UMLAggregation)))
-                                )
-                        } else if !rf_str.contains(r"*mut") && !rf_str.contains(r"*const") {
-                            get_paths_str_from_ast_node(rf)
-                                .iter()
-                                .for_each(|p| results.push(
-                                    UMLEntity::UMLRelation(UMLRelation::new(&self.name().unwrap().text().to_string(), &p, UMLRelationKind::UMLComposition)))
-                                )
-                        }
-                    },
-                    _ => ()
-                }
-            }
-        };
-
-        results
-    }
-}
-
-impl HasUMLEntity for ast::Impl {
-    fn get_uml_entities(&self) -> Vec<UMLEntity> {
-        let mut results = vec![];
-
-        // get struct name
-        let mut impl_fn_names = vec![];
-        let struct_name: String = strip_trait_bound(&self.self_ty().unwrap().to_string());
-
-        let mut dep_list: Vec<String> = vec![];
-        let mut asct_list: Vec<String> = vec![];
-        for node in self.syntax().descendants() {
-            match_ast! {
-                match node {
-                    // get impl functions' names
-                    ast::Fn(f) => {
-                        impl_fn_names.push(get_fn_full_name(&f));
-                    },
-                    // get Dependency and Association Relations
-                    ast::ParamList(pl) => {
-                        dep_list.append(&mut get_paths_str_from_ast_node(pl));
-                    },
-                    ast::BlockExpr(ex) => {
-                        dep_list.append(&mut get_paths_str_from_ast_node(ex));
-                    },
-                    ast::RetType(rt) => {
-                        asct_list.append(&mut get_paths_str_from_ast_node(rt));
-                    },
-                    _ => ()
-                }
-            }
-        }
-
-        // first add Association Relation, then add dependency relation if the name not occured in assocaitions
-        results.extend(
-            asct_list.iter().map(|p| UMLEntity::UMLRelation(UMLRelation::new(&p, &struct_name, UMLRelationKind::UMLAssociationUni)))
-        );
-        let mut dep_set: Vec<&String> = dep_list.iter().filter(|p| !asct_list.contains(p)).collect();
-        dep_set.sort();
-        dep_set.dedup();
-        results.extend(
-            dep_set.iter().map(|p| UMLEntity::UMLRelation(UMLRelation::new(&p, &struct_name, UMLRelationKind::UMLDependency)))
-        );
-
-
-        // get trait if there is any
-        match self.trait_() {
-            Some(tt) => {
-                results.push(
-                    UMLEntity::UMLRelation(UMLRelation::new(&struct_name, &strip_trait_bound(&tt.to_string()), UMLRelationKind::UMLRealization))
-                );
-                // println!("trait: {}", tt.to_string());
-            },
-            None => {
-                results.push(UMLEntity::UMLClass(UMLClass::new(&struct_name, vec![], impl_fn_names, UMLClassKind::UMLClass)));
-            }
-        }
-        
-
-        results
-    }
-}
-
-fn strip_trait_bound(s: &str) -> String {
-    let class_name: Vec<&str> = s.split(r"<").collect();
-    String::from(class_name[0])
-}
-
-fn get_paths_str_from_ast_node(node: impl ast::AstNode) -> Vec<String> {
-    // get raw relation string
-    let mut results = vec![];
-    for node in node.syntax().descendants() {
-        match_ast! {
-            match node {
-                ast::Path(p) => {
-                    results.push(p.to_string())
-                },
-                _ => ()
-            }
-        }
-        // println!("{:?}", node);
-        // println!("{}", node);
-    };
-    results
-}
-
-impl HasUMLEntity for ast::Fn {
-    fn get_uml_entities(&self) -> Vec<UMLEntity> {
-        let mut results: Vec<UMLEntity> = vec![];
-        let f_name = self.name().unwrap().text().to_string();
-        let full_name: String = get_fn_full_name(self);
-
-        // visit all Fn descendants and process CallExpr
-        for node in self.syntax().descendants() {
-            match_ast! {
-                match node {
-                    ast::CallExpr(it) => {
-                        let call_name = get_call_expr_fn_names(it);
-                        results.push(UMLEntity::UMLRelation(UMLRelation::new(&f_name, &call_name, UMLRelationKind::UMLDependency)))
-                    },
-                    _ => {
-                        // println!("{:?}", node);
-                        // println!("{}", node)
-                    },
-                }
-            }
-        }
-        results.push(UMLEntity::UMLFn(UMLFn::new(&f_name, &full_name)));
-        results
-    }
-}
-
-fn get_fn_full_name(f: &ast::Fn) -> String {
-    // include param list, return type
-    let f_name = f.name().unwrap().text().to_string();
-    let mut full_name: String = f_name.clone();
-
-    // visit all Fn descendants and process CallExpr
-    for node in f.syntax().descendants() {
-        match_ast! {
-            match node {
-                ast::ParamList(pl) => {
-                    full_name.push_str(&pl.to_string());
-                },
-                ast::RetType(rt) => {
-                    full_name.push_str(" ");
-                    full_name.push_str(&rt.to_string());
-                },
-                _ => {
-                    // println!("{:?}", node);
-                    // println!("{}", node)
-                },
-            }
-        }
-    }
-    full_name
-}
-
-fn get_call_expr_fn_names(call_exp: ast::CallExpr) -> String {
-    let call_expr = call_exp.to_string();
-    let call_names: Vec<&str> = call_expr.split("(").collect();
-    String::from(call_names[0])
-}
+use super::to_uml_entities::HasUMLEntity;
 
 pub struct AstParser;
